+++ tksheet/other_classes.py
@@ -17,14 +17,8 @@ Box_t = namedtuple(
     "Box_t",
     "from_r from_c upto_r upto_c type_",
 )
-Box_st = namedtuple(
-    "Box_st",
-    "coords type_"
-)
-Loc = namedtuple(
-    "Loc",
-    "row column"
-)
+Box_st = namedtuple("Box_st", "coords type_")
+Loc = namedtuple("Loc", "row column")
 
 Highlight = namedtuple(
     "Highlight",
@@ -476,4 +470,3 @@ Selected = namedtuple(
         None,
     ),
 )
-