+++ tksheet/row_index.py
@@ -31,10 +31,10 @@ from .functions import (
     consecutive_chunks,
     ev_stack_dict,
     event_dict,
-    rounded_box_coords,
     get_n2a,
     is_contiguous,
     num2alpha,
+    rounded_box_coords,
     try_binding,
 )
 from .other_classes import (
@@ -965,7 +965,7 @@ class RowIndex(tk.Canvas):
             y1,
             x2,
             y2,
-            radius=9 if self.PAR.ops.rounded_boxes else 0,
+            radius=8 if self.PAR.ops.rounded_boxes else 0,
         )
         if isinstance(iid, int):
             self.coords(iid, coords)